perf(spectrum): carry per-band dB at f32, not f64
P-7 in the protocol round. The payload's dB rows were the DSP's f64 all the way to
the webview, which then packed them into an Int16 centi-dB slab -- so the wire was
carrying eight bytes to deliver a value with a 0.01 dB destination. Narrowing at the
payload boundary halves the section and shrinks the JSON path with it; a
production-width frame is now 36,000 bytes against the 131,886 this round started
from.

The narrowing is a precision change, so it comes with the pairing the design doc
asked for rather than an argument. Swept across -200..200 dB, the worst f64-to-f32
error is 7.32e-6 dB -- 1366 times finer than the centi-dB step the history slab
stores, four orders under the 0.1 dB the panel shows. A second test holds the
property that matters more than the magnitude: two values a centi-dB apart cannot
collapse onto one f32, so nothing the display can distinguish can be merged by the
wire.

Doing this after the transport landed rather than with it was deliberate. The f64
step moved bytes while changing no value; this one changes values while moving no
structure. If the real window turns up a problem, which of the two caused it is not
a question.

+++ src-tauri/src/ipc/types.rs
@@ -60,11 +60,15 @@ pub struct AnalysisRequests {
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
+/// Per-band dB for one request key. The DSP computes these in `f64`, but they are narrowed here:
+/// the frontend's only destination for them is an Int16 centi-dB slab, and `f32` sits three orders
+/// of magnitude below that step (`spectrum_db_narrowing_stays_far_below_display_precision`). The
+/// narrowing halves what the wire carries, which is the point -- see `docs/working/perf/protocol.md`.
 pub struct SpectrumFrameResult {
-  pub smooth_db: Vec<f64>,
-  pub peak_db: Vec<f64>,
-  pub smooth_db_b: Vec<f64>,
-  pub peak_db_b: Vec<f64>,
+  pub smooth_db: Vec<f32>,
+  pub peak_db: Vec<f32>,
+  pub smooth_db_b: Vec<f32>,
+  pub peak_db_b: Vec<f32>,
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
@@ -190,9 +194,9 @@ pub struct MeterHistoryEntry {
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct SpectrumVisualEntry {
-  pub smooth_db: Vec<f64>,
+  pub smooth_db: Vec<f32>,
   /// Secondary smoothed per-band dB (empty unless view is lr/ms).
-  pub smooth_db_b: Vec<f64>,
+  pub smooth_db_b: Vec<f32>,
 }
 
 /// Per-request-key vectorscope sample for one visual history tick.
